Restructure repository files into package folders (src/controllers, src/simulation, src/visualization, src/evaluation)
