+++ .mypy.ini
@@ -1,3 +1,3 @@
 [mypy]
 ignore_missing_imports = True
-exclude = venv|src
+exclude = venv|src/main.py
