+++ tests/unit/test_data_logger.py
@@ -17,8 +17,8 @@ from pathlib import Path
 
 import pytest
 
-from data_logger import (ComparisonAnalyzer, CSVExporter, DataLogger,
-                         JSONExporter)
+from src.evaluation.data_logger import (ComparisonAnalyzer, CSVExporter,
+                                        DataLogger, JSONExporter)
 
 # ---------------------------------------------------------------------------
 # Helpers
